Fix how flush_pos is compared to commit_lsn.
When we logically replicate the message with LSN equal to exactly X,
we need to wait for `flush_pos` of other replicas to become strictly
greater than X, not greater or equal.

Both `restart_lsn` and `confirmed_flush` represent the oldest LSN
that must still be kept so it can still be replicated. That means
that it might NOT yet be replicated.